Added sync
This commit includes a new ansible playbook for copying images up to the server
using rsync: `ansible-playbook sync`. The manifests directory and image
directory on the server is now configurable in group_vars/all/config.yml

In addition the `inventory_hostname` variable is used rather than requiring the
user to put the hostname in two places.